+++ src/app/modules/color-scheme/color-bar/color-bar.component.html
@@ -1,16 +1,11 @@
 <div *ngIf="colorScheme.type === 'discrete'; else gradient">
-  <div *ngFor="color of colorScheme.colors; index as i" 
-    class="{{colorScheme.type}}" (click)="colorChanged(i)" 
-    [class.currentcolor]="isColorSelected(i)" 
-    [style.background-color]="getColor(i)"
+  <div *ngFor="color of colorScheme.colors; index as i" class="{{colorScheme.type}}"
+    [class.currentcolor]="isColorSelected(i)" (click)="colorChanged(i)" [style.background-color]="getColor(i)"
     [style.background-image]="gradientstyle">
   </div>
 </div>
 
 <ng-template #gradient>
-  <div
-    class="{{colorScheme.type}}"
-    [style.background-image]="gradientstyle">
+  <div class="{{colorScheme.type}}" [style.background-image]="gradientstyle">
   </div>
 </ng-template>
-
