+++ src/app/modules/color-scheme/color-bar/color-bar.component.ts
@@ -1,5 +1,6 @@
-import { Component, Input, Output, EventEmitter, OnInit } from '@angular/core';
-import { ColorScheme } from '../color-scheme-popup/color-scheme-popup.component';
+import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
+
+import { ColorScheme } from '../color-schemes';
 
 /**
  * Color bar component for the color scheme popup and color scheme selector
@@ -16,31 +17,33 @@ export class ColorBarComponent implements OnInit {
    */
   @Input() colorScheme: ColorScheme;
 
+  @Input() selected = false;
+
   /**
    * Emits the newly selected color
    */
   @Output() colorChange = new EventEmitter<string | undefined>();
 
   /**
-   * Array for iterating through colors to be displayed
-   */
-  colorIndex: number[];
-
-  /**
    * Enables dynamic styling for gradient bars
    */
   gradientstyle: string;
 
-  @Input() selected = false;
-
   selectedColorIndex = 0;
 
   /**
-   * Sets the color status array, color index array, and gradient style (when applicable) on load
+   * Sets the gradient style (when applicable) on load
    */
   ngOnInit() {
-    this.colorIndex = this.colorScheme.type === 'discrete' ? [0,1,2,3,4,5,6] : [0];
-    this.gradientstyle = this.colorScheme.type === 'gradient' ? `linear-gradient(to right, ${this.getColor(0)} , ${this.getColor(1)}, ${this.getColor(2)})` : 'none';
+    this.gradientstyle = this.colorScheme.type === 'gradient' ? `linear-gradient(to right, ${this.gradientColorString()})` : 'none';
+  }
+
+  gradientColorString() {
+    const result = [];
+    for (let i = 0; i < this.colorScheme.colors.length; i++) {
+      result.push(this.colorScheme.colors[i] + ' ' + this.colorScheme.positions[i] * 100 + '%');
+    }
+    return result.join(', ');
   }
 
   /**
